ATM_Action_Recognition/action_requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

# --- Core ML/Data Libraries ---
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
numpy
scikit-learn # For train/test split
pillow

# --- Dataset & Utilities ---
kagglehub
python-dotenv
tqdm

# --- Video/Image Handling ---
opencv-python